% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer75indices}
\alias{cancer75indices}
\title{hash tables underlying metadata search for htxapp}
\format{list of environments}
\usage{
cancer75indices
}
\description{
hash tables underlying metadata search for htxapp
}
\keyword{datasets}
