% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2envs.R
\name{csv2envs}
\alias{csv2envs}
\title{utility to create hashtables (environments) linking study/experiment/keyword
for use in searching metadata related to HumanTranscriptomeCompendium elements}
\usage{
csv2envs(csv, envset = NULL, expenv = new.env(hash = TRUE),
  studenv = new.env(hash = TRUE), cleanFields = list(".*id$", ".name$",
  "_name$"))
}
\arguments{
\item{csv}{character(1) path to a CSV-formated file assumed to have fields study.accession, experiment.accession}

\item{envset}{a list of environments with elements expenv, studenv, kwexenv, kwstenv}

\item{expenv}{an environment}

\item{studenv}{an environment}

\item{cleanFields}{a list with regular expressions as elements; fields matching elements of the
list will not contributes tokens for searching}
}
\description{
utility to create hashtables (environments) linking study/experiment/keyword
for use in searching metadata related to HumanTranscriptomeCompendium elements
}
\note{
Purpose is to create and add content to maps between tokens in metadata and the studies and experiments
that bear them.  "Stopwords" are removed.
}
