% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_trace.R
\name{feature_trace}
\alias{feature_trace}
\title{plot traces for the 'paired' or 'tripled' design}
\usage{
feature_trace(featuresym, se, tx = function(x) log(x + 1),
  id = "patient.id", strat = "source_name")
}
\arguments{
\item{featuresym}{character(1) name of an assay (rowname of SummarizedExperiment)}

\item{se}{SummarizedExperiment instance}

\item{tx}{function that transforms numerical input to desired scale for plotting}

\item{id}{character(1) name of subject identifier, a vector in colData that
partitions its rows into the collections of contributions from each subject}

\item{strat}{character(1) name of variable in colData that gives the
type of each of the repeated measures}
}
\description{
plot traces for the 'paired' or 'tripled' design
}
\examples{
data(hccLinc4)
requireNamespace("ggplot2")
feature_trace("TERC", hccLinc4) + xlab(" ") + guides(colour="none")
}
