% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hccLinc4}
\alias{hccLinc4}
\title{small SummarizedExperiment with repeated measures}
\format{RangedSummarizedExperiment instance}
\usage{
hccLinc4
}
\description{
small SummarizedExperiment with repeated measures
}
\note{
There are 4 features, 58 individuals, derived from data
related to PMID 28194035.
}
\keyword{datasets}
