% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n1}
\alias{n1}
\title{hash tables underlying metadata search for htxapp}
\format{list of environments}
\usage{
n1
}
\description{
hash tables underlying metadata search for htxapp
}
\note{
This was produced for a small proof of concept.
}
\keyword{datasets}
