% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleBindMetadata.R
\name{simpleBindMetadata}
\alias{simpleBindMetadata}
\title{for an htxapp download involving only one study, bind the sample.attributes information
to the colData}
\usage{
simpleBindMetadata(se)
}
\arguments{
\item{se}{a SummarizedExperiment instance}
}
\description{
for an htxapp download involving only one study, bind the sample.attributes information
to the colData
}
\note{
htxapp returns a SummarizedExperiment instance with additional data.frames in the
metadata() list.  For the case of one study returned, there is only one such data.frame
and it is clear how to combine the information in it with the metadata already in colData.
When multiple studies are selected, there will likely be different field sets in each
sample.attributes component, so build a unified colData requires manual work.
}
