% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studyTitles}
\alias{studyTitles}
\title{vector of titles of studies available}
\format{character vector}
\usage{
studyTitles
}
\description{
vector of titles of studies available
}
\keyword{datasets}
